NCS Player initial commit
I've got a proto type working. I talk about this a little i the
README.md but I'll re-itterate here. This app is meant to be a
proving grounds for my idea of making a video player with kivy
for independent content creators, project Vapaa Aani. I decided
before I delve to deeply into that project, that I would fist go
about the process of making a simpler video player app.

Right now it takes a static list of videos and allows the user to
pick one and play it. My goal is to have the whole video list from
the NCS channel available, but I decided to cross the bridge when I
needed to and for now just have a list of videos.

Right now you also cannot change video after you have selected one.
That is a bit bother some but I'll sort it out next. The important
thing is that the prototype plays videos, that was the challenge.