NCS Player app works.
I've got the NCS player up and running on my tablet. There is
always a few little things that are broken. So after a little bit
of tweaking it finally spun up on iron. PROOF OF CONCEPT! I now
will have to spend some time figuring out how to clean it up and
get the functionality where it needs to be.

While this is just a demo app I do plan on taking it the whole way.
Mostly as an exercise. I want to go through all the steps of making
a polished app and submitting it to the app store. I'm also
considering using this app as an experiment in ads. I know that I
previously didn't want to do ad supported, and I still don't want
to do it, but I feel like I should at least figure it out.